Add quotes around example account numbers
The account numbers in the example don't require quotes but if a user
were to fill in account numbers with leading zeroes then YAML won't
interpret those as octal (for YAML 1.1) or base 10 integer (YAML 1.2).
This simple change makes it less likely to run into this issue when
using this template as a starter.